Moved command shell creation across to winrm_login, rather than winrm_cmd
