+++ documentation/modules/auxiliary/scanner/winrm/winrm_cmd.md
@@ -1,6 +1,5 @@
 ## Description
 This module creates a WinRM command shell. It needs login credentials to do so.
-The module launches powershell.exe and communicates with it over stdin/stdout/stderr.
 
 ## Verification Steps
 
@@ -8,13 +7,12 @@ The module launches powershell.exe and communicates with it over stdin/stdout/st
 1. Do: ```set RHOSTS [IP]```
 1. Do: ```set USERNAME [USERNAME]```
 1. Do: ```set PASSWORD [PASSWORD]```
-1. Optionally Do: ```set CreateSession false```
-1. Optionally Do: ```set CMD [WINDOWS COMMAND]```
 1. Do: ```run```
 
 ## Scenarios
 
-### Create shell
+
+### Run single command
 
 ```
 msf6 > use scanner/winrm/winrm_cmd
@@ -24,43 +22,6 @@ msf6 auxiliary(scanner/winrm/winrm_cmd) > set password pass12345
 password => pass12345
 msf6 auxiliary(scanner/winrm/winrm_cmd) > set rhosts 192.168.1.205
 rhosts => 192.168.1.205
-msf6 auxiliary(scanner/winrm/winrm_cmd) > run
-[*] Command shell session 1 opened (WinRM) at 2021-09-02 20:26:41 +1000
-
-[*] Scanned 1 of 1 hosts (100% complete)
-[*] Auxiliary module execution completed
-msf6 auxiliary(scanner/winrm/winrm_cmd) > sessions 
-
-Active sessions
-===============
-
-  Id  Name  Type           Information                                         Connection
-  --  ----  ----           -----------                                         ----------
-  1         shell windows  WinRM Administrator:pass12345 (VM01\Administrator)  WinRM (192.168.1.205)
-
-msf6 auxiliary(scanner/winrm/winrm_cmd) > sessions 1
-[*] Starting interaction with 1...
-
-Windows PowerShell
-Copyright (C) Microsoft Corporation. All rights reserved.
-
-Try the new cross-platform PowerShell https://aka.ms/pscore6
-
-PS C:\Users\Administrator>
-```
-
-### Run single command (no shell)
-
-```
-msf6 > use scanner/winrm/winrm_cmd
-msf6 auxiliary(scanner/winrm/winrm_cmd) > set username Administrator
-username => Administrator
-msf6 auxiliary(scanner/winrm/winrm_cmd) > set password pass12345
-password => pass12345
-msf6 auxiliary(scanner/winrm/winrm_cmd) > set rhosts 192.168.1.205
-rhosts => 192.168.1.205
-msf6 auxiliary(scanner/winrm/winrm_cmd) > set CreateSession false
-CreateSession => false
 msf6 auxiliary(scanner/winrm/winrm_cmd) > set cmd whoami /priv
 cmd => whoami /priv
 msf6 auxiliary(scanner/winrm/winrm_cmd) > run
